requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fonttools==4.34.3
idna==3.3
imageio==2.19.3
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
networkx==2.8.4
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
stack-data==0.3.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
